fix cpp warnings

+++ mettagrid/benchmarks/test_mettagrid_env_benchmark.cpp
@@ -27,7 +27,7 @@ namespace py = pybind11;
 // We'll know we've succeeded when this file has zero references to pybind11!
 
 // Helper functions for creating configuration and map
-GameConfig CreateBenchmarkConfig(int num_agents) {
+GameConfig CreateBenchmarkConfig(unsigned int num_agents) {
   std::vector<std::string> inventory_item_names = {"ore", "heart"};
 
   std::shared_ptr<ActionConfig> action_cfg = std::make_shared<ActionConfig>(
@@ -80,7 +80,7 @@ GameConfig CreateBenchmarkConfig(int num_agents) {
   return GameConfig(num_agents, 10000, 11, 11, inventory_item_names, 100, actions_cfg, objects_cfg);
 }
 
-py::list CreateDefaultMap(int num_agents_per_team = 2) {
+py::list CreateDefaultMap(size_t num_agents_per_team = 2) {
   py::list map;
   const int width = 32;
   const int height = 32;
@@ -99,7 +99,7 @@ py::list CreateDefaultMap(int num_agents_per_team = 2) {
   }
 
   // Place agents symmetrically
-  int agents_placed = 0;
+  size_t agents_placed = 0;
   std::vector<std::pair<int, int>> positions = {{8, 8},   {8, 24},  {24, 8},  {24, 24}, {16, 8},  {16, 24}, {8, 16},
                                                 {24, 16}, {12, 12}, {12, 20}, {20, 12}, {20, 20}, {16, 16}, {16, 12},
                                                 {16, 20}, {12, 16}, {20, 16}, {10, 10}, {10, 22}, {22, 10}, {22, 22},
@@ -108,8 +108,8 @@ py::list CreateDefaultMap(int num_agents_per_team = 2) {
   for (size_t i = 0; i < positions.size() && agents_placed < num_agents_per_team * 2; ++i) {
     auto [r, c] = positions[i];
     std::string team = (agents_placed % 2 == 0) ? "agent.team1" : "agent.team2";
-    py::list row = map[r].cast<py::list>();
-    row[c] = team;
+    py::list row = map[static_cast<size_t>(r)].cast<py::list>();
+    row[static_cast<size_t>(c)] = team;
     agents_placed++;
   }
 
@@ -117,10 +117,10 @@ py::list CreateDefaultMap(int num_agents_per_team = 2) {
 }
 
 // Utility function to generate valid random actions
-py::array_t<int> GenerateValidRandomActions(MettaGrid* env, int num_agents, std::mt19937* gen) {
+py::array_t<int> GenerateValidRandomActions(MettaGrid* env, size_t num_agents, std::mt19937* gen) {
   // Get the maximum argument values for each action type
   py::list max_args = env->max_action_args();
-  int num_actions = py::len(env->action_names());
+  size_t num_actions = py::len(env->action_names());  // Change to size_t
 
   // Create actions array
   std::vector<py::ssize_t> shape = {static_cast<py::ssize_t>(num_agents), 2};
@@ -128,14 +128,14 @@ py::array_t<int> GenerateValidRandomActions(MettaGrid* env, int num_agents, std:
   auto actions_ptr = static_cast<int*>(actions.request().ptr);
 
   // Initialize distributions
-  std::uniform_int_distribution<> action_dist(0, num_actions - 1);
+  std::uniform_int_distribution<> action_dist(0, static_cast<int>(num_actions) - 1);
 
-  for (int i = 0; i < num_agents; ++i) {
+  for (size_t i = 0; i < num_agents; ++i) {
     // Choose random action type
     int action_type = action_dist(*gen);
 
     // Get max allowed argument for this action type
-    int max_arg = py::cast<int>(max_args[action_type]);
+    int max_arg = py::cast<int>(max_args[static_cast<size_t>(action_type)]);
 
     // Choose random valid argument (0 to max_arg inclusive)
     int action_arg = 0;
@@ -153,14 +153,14 @@ py::array_t<int> GenerateValidRandomActions(MettaGrid* env, int num_agents, std:
 }
 
 // Pre-generate a sequence of actions for the benchmark
-std::vector<py::array_t<int>> PreGenerateActionSequence(MettaGrid* env, int num_agents, int sequence_length) {
+std::vector<py::array_t<int>> PreGenerateActionSequence(MettaGrid* env, size_t num_agents, size_t sequence_length) {
   std::vector<py::array_t<int>> action_sequence;
-  action_sequence.reserve(sequence_length);
+  action_sequence.reserve(static_cast<size_t>(sequence_length));
 
   // Use a deterministic seed for reproducible benchmarks
   std::mt19937 gen(42);
 
-  for (int i = 0; i < sequence_length; ++i) {
+  for (size_t i = 0; i < sequence_length; ++i) {
     action_sequence.push_back(GenerateValidRandomActions(env, num_agents, &gen));
   }
 
@@ -170,7 +170,7 @@ std::vector<py::array_t<int>> PreGenerateActionSequence(MettaGrid* env, int num_
 // Matching Python test_step_performance_no_reset
 static void BM_MettaGridStep(benchmark::State& state) {  // NOLINT(runtime/references)
   // Setup with default 4 agents (matching Python benchmark config)
-  int num_agents = 4;
+  unsigned int num_agents = 4;
   auto cfg = CreateBenchmarkConfig(num_agents);
   auto map = CreateDefaultMap(2);
 
@@ -178,7 +178,7 @@ static void BM_MettaGridStep(benchmark::State& state) {  // NOLINT(runtime/refer
   env->reset();
 
   // Verify agent count
-  if (env->num_agents() != num_agents) {
+  if (static_cast<unsigned int>(env->num_agents()) != num_agents) {
     state.SkipWithError("Agent count mismatch");
     return;
   }
